+++ test/MinimalBinding/MinimalBinding.cpp
@@ -29,14 +29,13 @@ int _tmain(int argc, _TCHAR* argv[])
 
 	Flathead *pFH = new Flathead(cfg);
 	char buffer[4096];
-
-	WriteToFile("lib/HotReload.js", "this.myVar = false; var that = this; exports.value = this.myVar; module.unload = function () { that.myVar = true; }");
+	WriteToFile("lib/HotReload.js", "var global = this; exports.value = this.myVar; module.unload = function () { global.unloaded = true; }");
 
 	pFH->Execute("require('./HotReload').value;", buffer);
 
 	std::this_thread::sleep_for(std::chrono::seconds(1)); // Only limitation is the resolution reported by filesystem =\
 
-	WriteToFile("lib/HotReload.js", "console.log(JSON.stringify(this)); exports.value = this.myVar;");
+	WriteToFile("lib/HotReload.js", "exports.value = this.unloaded;");
 
 	pFH->Tick(1.0f);