+++ changelog.md
@@ -6,6 +6,10 @@ For more information about how to maintain this document, [please keep a changel
 
 ## [Unreleased][unreleased]
 
+### Added
+* Module hot swapping 
+* Provided unload event for hot swapping
+
 ## [2.1.1][2.1.1]
 
 ### Added